Downloads/SWT-FSE-develop/requirements.txt
aniso8601==1.1.0
Flask==0.10.1
Flask-RESTful==0.3.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
numpy==1.11.0
PyMySQL==0.7.3
python-dateutil==2.5.3
pytz==2016.4
six==1.10.0
tornado==4.3
Werkzeug==0.11.9
xlwt
#git+ssh://git@github.com/lfcunha/flask-restful.git
https://github.com/lfcunha/flask-restful.git
boto
boto3
falcon
gevent
requests
pyjwt
bson
pymongo
pytest
pytest-flask